feat: 补充V146下GroupInfo错乱的升级修复代码 (#1297)
:( 先行使特权了
